+++ packages/flutter_riverpod/lib/flutter_riverpod.dart
@@ -1,5 +1,9 @@
-// ignore: invalid_export_of_internal_element, Already tackled by riverpod/riverpod.dart. If we export internals, that's on purpose.
-export 'package:riverpod/riverpod.dart';
+// ignore: invalid_export_of_internal_element, those are marked with internal. We export them for backward compatibility
+export 'package:riverpod/src/riverpod_internals_export.dart';
+export 'package:riverpod/src/riverpod_without_legacy.dart';
 
-export 'src/core.dart' hide ProviderScopeState;
-export 'src/providers/legacy/change_notifier_provider.dart';
+@Deprecated(
+  'This is old API. Use `package:riverpod/legacy.dart` if you want to keep using it.',
+)
+export './legacy.dart';
+export 'src/riverpod_without_legacy.dart';