+++ packages/flutter_riverpod/lib/src/riverpod_internals_export.dart
@@ -1,0 +1,7 @@
+export 'package:riverpod/src/riverpod_internals_export.dart';
+
+// ignore: invalid_export_of_internal_element, For the sake of backward compatibility. Remove in 3.0
+export 'providers/legacy/change_notifier_provider.dart'
+    show
+        ChangeNotifierProviderElement,
+        AutoDisposeChangeNotifierProviderElement;