+++ packages/flutter_riverpod/lib/src/riverpod_without_legacy.dart
@@ -1,0 +1,3 @@
+export 'package:riverpod/src/riverpod_without_legacy.dart';
+
+export 'core.dart' hide ProviderScopeState;