+++ packages/riverpod/test/old/legacy_providers/deprecated/state_provider/state_provider_family_test.dart
@@ -1,4 +1,5 @@
 import 'package:riverpod/riverpod.dart';
+import 'package:riverpod/legacy.dart';
 import 'package:test/test.dart';
 
 import '../../../utils.dart';