+++ packages/riverpod/test/old/legacy_providers/stream_provider/auto_dispose_stream_provider_test.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:mockito/mockito.dart';
 import 'package:riverpod/riverpod.dart';
+import 'package:riverpod/legacy.dart';
 import 'package:test/test.dart';
 
 import '../../utils.dart';