+++ packages/riverpod/test/providers/provider/provider_test.dart
@@ -2,6 +2,7 @@
 
 import 'package:mockito/mockito.dart';
 import 'package:riverpod/riverpod.dart';
+import 'package:riverpod/legacy.dart';
 import 'package:test/test.dart';
 
 import '../../utils.dart';