+++ packages/riverpod/test/src/providers/notifier_test.dart
@@ -3,6 +3,7 @@
 import 'package:meta/meta.dart';
 import 'package:mockito/mockito.dart';
 import 'package:riverpod/riverpod.dart' hide ErrorListener;
+import 'package:riverpod/legacy.dart';
 import 'package:test/test.dart';
 
 import '../../utils.dart';