+++ website/docs/concepts/combining_provider_states/filtered_todo_list_provider/raw.dart
@@ -1,4 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_riverpod/legacy.dart';
+
 import '../todo_list_provider/raw.dart';
 
 enum Filter {