+++ website/static/snippets/combine.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: omit_local_variable_types
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_riverpod/legacy.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'combine.g.dart';