+++ packages/flutter_riverpod/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased minor
 
+- `StateNotifierProvider/StateProvider/ChangeNotifierProvider` and relevant APIs are considered deprecated.
+  To keep using them, import `package:flutter_riverpod/legacy.dart'`.
 - Added various experimental features:
   - A new syntax for providers
   - Mutations